==> 12_plots of most parsimonious model.R <==
rm(list=ls())

library(dplyr)
library(tidyr)
library(gridExtra)
library(grid)
library(GlobalArchive)
library(stringr)
library(ggplot2)
library(gamm4)

# Set the study name
name <- '2020_south-west_stereo-BRUVs' # for the study

## Set working directory----
working.dir <- dirname(rstudioapi::getActiveDocumentContext()$path)

## Set sub directories----
d.dir <- paste(working.dir,"Data/Tidy",sep="/") 
h.dir <- paste(working.dir, "Data/Habitat/BRUV Style annotation/tidy data",sep="/") 
s.dir <- paste(working.dir,"shapefiles",sep="/")
p.dir <- paste(working.dir,"Plots",sep="/")
m.dir <- paste(working.dir,"Model Out GAM", sep="/")

# Bring in and format the data----
setwd(d.dir)
dir()

# plots of the most parsimonious models----
# Theme-
Theme1 <-
  theme( # use theme_get() to see available options
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    # legend.background = element_rect(fill="white"),
    legend.background = element_blank(),
    legend.key = element_blank(), # switch off the rectangle around symbols in the legend
    legend.text = element_text(size=15),
    legend.title = element_blank(),
    legend.position = c(0.2, 0.8),
    text=element_text(size=15),
    strip.text.y = element_text(size = 15,angle = 0),
    axis.title.x=element_text(vjust=0.3, size=15),
    axis.title.y=element_text(vjust=0.6, angle=90, size=15),
    axis.text.x=element_text(size=15),
    axis.text.y=element_text(size=15),
    axis.line.x=element_line(colour="black", size=0.5,linetype='solid'),
    axis.line.y=element_line(colour="black", size=0.5,linetype='solid'),
    strip.background = element_blank())


# Bring in and format the raw data----

# Load the dataset -
# MaxN ----
maxn <-read.csv(paste(name, 'complete.maxn.csv',sep=".")) %>%
  dplyr::select(campaignid, sample, scientific, maxn, family, genus, species) %>%
  dplyr::mutate(sample=str_replace_all(.$sample,c("FHC01"="FHCO1","FHC02"="FHCO2","FHC03"="FHCO3"))) %>%
  dplyr::glimpse()

# Metadata ----
metadata <- read.csv(paste(name, 'checked.metadata.csv',sep=".")) %>%
  dplyr::mutate(status = as.factor(status)) %>%
  dplyr::mutate(sample = as.factor(sample)) %>%
  dplyr::mutate(planned.or.exploratory = as.factor(planned.or.exploratory)) %>%
  dplyr::mutate(site = as.factor(site)) %>%
  dplyr::filter(successful.count%in%c("Yes")) %>%
  dplyr::mutate(sample=str_replace_all(.$sample,c("FHC01"="FHCO1","FHC02"="FHCO2","FHC03"="FHCO3"))) %>%
  dplyr::glimpse()

# Bathymetry derivatives ----
bathy <- read.csv(paste(name, 'bathymetry.derivatives.csv',sep=".")) %>%
  dplyr::mutate(sample=str_replace_all(.$sample,c("FHC01"="FHCO1","FHC02"="FHCO2","FHC03"="FHCO3"))) %>%
  dplyr::glimpse()

# Distance to boat ramp ----
ramps <- read.csv(paste(name, 'distance.to.ramp.csv',sep=".")) %>%
  dplyr::mutate(sample=str_replace_all(.$sample,c("FHC01"="FHCO1","FHC02"="FHCO2","FHC03"="FHCO3"))) %>%
  dplyr::glimpse()

test<-left_join(metadata.fh,ramps)

ggmod.distance.to.ramp <- ggplot() +
  #ylab("")+
  #xlab("Distance to ramp")+
  #scale_color_manual(labels = c("Fished", "SZ"),values=c("red", "black"))+
  geom_point(data=test,aes(x=longitude,y=latitude, col=distance.to.ramp),  alpha=0.75, size=2,show.legend=TRUE)#+
  ##geom_line(data=predicts.pj.distance.to.ramp,aes(x=distance.to.ramp,y=maxn),alpha=0.5)+
  #geom_line(data=predicts.pj.distance.to.ramp,aes(x=distance.to.ramp,y=maxn - se.fit),linetype="dashed",alpha=0.5)+
  #geom_line(data=predicts.pj.distance.to.ramp,aes(x=distance.to.ramp,y=maxn + se.fit),linetype="dashed",alpha=0.5)+
 # theme_classic()+
 # Theme1

# Habitat ----
habitat.2020.10 <- read.csv("2020-10_south-west_stereo-BRUVs_BRUV_style.broad.habitat.csv") %>%
  dplyr::select(-c(rowid.x,rowid.y)) %>%
  dplyr::mutate(campaignid = "2020-10_south-west_stereo-BRUVs") %>%
  dplyr::glimpse()

summary(habitat.2020.10)

habitat.2020.06 <- read.csv("2020-06._broad.habitat_BRUV_Style.csv") %>%
  dplyr::select(-c(latitude,longitude,date,time,site,location,successful.count)) %>%
  dplyr::mutate(campaignid = "2020-06_south-west_stereo-BRUVs") %>%
  dplyr::glimpse()

summary(habitat.2020.06) # 0-100

habitat <-bind_rows(habitat.2020.06, habitat.2020.10) %>%
  tidyr::replace_na(list(broad.Consolidated=0,
                         broad.Macroalgae=0,
                         broad.Seagrasses=0,
                         broad.Sponges=0,
                         broad.Unconsolidated=0,
                         broad.Bryozoa=0,
                         broad.Hydroids=0,
                         broad.Octocoral.Black=0,
                         broad.Stony.corals=0,
                         fov.Facing.Up=0)) %>%
  ga.clean.names() %>%
  dplyr::mutate(broad.reef = broad.bryozoa + broad.consolidated + broad.hydroids + broad.macroalgae + broad.octocoral.black + broad.seagrasses + broad.sponges + broad.stony.corals) %>%
  dplyr::select(order(colnames(.))) %>%
  dplyr::select(campaignid,sample,everything()) %>% # re-ordering hab columns 
  dplyr::mutate(sample=str_replace_all(.$sample,c("FHC01"="FHCO1","FHC02"="FHCO2","FHC03"="FHCO3"))) %>%
  dplyr::glimpse()

# Create total abundance and species richness ----
ta.sr <- maxn %>%
  dplyr::ungroup() %>%
  dplyr::group_by(scientific,sample) %>%
  dplyr::summarise(maxn = sum(maxn)) %>%
  tidyr::spread(scientific,maxn, fill = 0) %>%
  dplyr::mutate(total.abundance=rowSums(.[,2:(ncol(.))],na.rm = TRUE )) %>% #Add in Totals
  dplyr::mutate(species.richness=rowSums(.[,2:(ncol(.))] > 0)) %>% # double check these
  dplyr::select(sample,total.abundance,species.richness) %>%
  tidyr::gather(.,"scientific","maxn",2:3) %>%
  dplyr::glimpse()

# Create abundance of all recreational fished species ----
setwd(d.dir)
dir()

# Select species of interest to model ----
species.maxn <- maxn %>%
  dplyr::filter(scientific %in% c("Sparidae Chrysophrys auratus",
                                  "Labridae Coris auricularis",
                                  "Heterodontidae Heterodontus portusjacksoni",
                                  "Monacanthidae Nelusetta ayraud"
  ))%>%
  dplyr::select(sample,scientific,maxn) %>%
  distinct()

# Split metadata into Fishing HWY and In/Out dataframes
summary(metadata)

metadata.fh <- metadata %>%
  dplyr::filter(depth<50)

metadata.io <- metadata %>%
  dplyr::filter(latitude<=(-33.96))

## Combine all the maxn data to be modeled into a single data frame
combined.maxn <- bind_rows(species.maxn, 
                           ta.sr)%>%
  left_join(metadata) %>%
  left_join(bathy) %>%
  left_join(ramps) %>%
  left_join(habitat) %>%
  distinct()

maxn.fh <- combined.maxn %>%
  semi_join(., metadata.fh) %>%
  filter(!scientific%in%c("Monacanthidae Nelusetta ayraud"))

maxn.io <- combined.maxn %>%
  semi_join(., metadata.io)

maxn.fh <- maxn.fh %>%
  dplyr::mutate(log.sponges = log(broad.sponges + 1)) %>%
  dplyr::mutate(log.tpi = log(tpi + 2)) %>%
  dplyr::mutate(log.roughness = log(roughness + 1)) %>%
  dplyr::mutate(log.slope = log(slope + 1))

maxn.io <- maxn.io %>%
  dplyr::mutate(log.sponges = log(broad.sponges + 1)) %>%
  dplyr::mutate(log.tpi = log(tpi + 2)) %>%
  dplyr::mutate(log.roughness = log(roughness + 1)) %>%
  dplyr::mutate(log.slope = log(slope + 1))

# Manually make the most parsimonious GAM models for each taxa ----
#### FISHING HWY ####
dat <- maxn.fh

unique(dat$scientific)

# MODEL Total abundance (log.tpi.by.status+mean.relief.by.status+status) ----
dat.ta <- dat %>% filter(scientific=="total.abundance")

gamm=gam(maxn~s(log.tpi,k=3,bs='cr', by=status)+s(mean.relief,k=3,bs='cr', by=status)+ s(site,bs="re") + status, family=poisson,data=dat.ta)

# predict - status ----
mod<-gamm
testdata <- expand.grid(mean.relief=mean(mod$model$mean.relief),
                        log.tpi=mean(mod$model$log.tpi),
                        site=(mod$model$site),
                        status = c("Fished","No-take"))%>%
  distinct()%>%
  glimpse()


fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.ta.status = testdata%>%data.frame(fits)%>%
  group_by(status)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# predict - log.tpi.by.status ----
mod<-gamm
testdata <- expand.grid(log.tpi=seq(min(dat$log.tpi),max(dat$log.tpi),length.out = 20),
                        mean.relief=mean(mod$model$mean.relief),
                        site=(mod$model$site),
                        status = c("Fished","No-take"))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.ta.log.tpi.x.status = testdata%>%data.frame(fits)%>%
  group_by(log.tpi,status)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# predict - mean.relief.by.status ----
mod<-gamm
testdata <- expand.grid(mean.relief=seq(min(dat$mean.relief),max(dat$mean.relief),length.out = 20),
                        log.tpi=mean(mod$model$log.tpi),
                        site=(mod$model$site),
                        status = c("Fished","No-take"))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.ta.mean.relief.x.status = testdata%>%data.frame(fits)%>%
  group_by(mean.relief,status)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()


# PLOTS for Total abundance ----
# status ----
ggmod.ta.status<- ggplot(aes(x=status,y=maxn,fill=status,colour=status), data=predicts.ta.status) +
  ylab(" ")+
  xlab('Status')+
  scale_fill_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  scale_colour_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  scale_x_discrete(limits = rev(levels(predicts.ta.status$status)))+
  geom_bar(stat = "identity")+
  geom_errorbar(aes(ymin = maxn-se.fit,ymax = maxn+se.fit),width = 0.5) +
  theme_classic()+
  Theme1+
  annotate("text", x = -Inf, y=Inf, label = "Total abundance",vjust = 1, hjust = -.1,size=5,fontface="italic")
ggmod.ta.status

# log.tpi.by.status ----
ggmod.ta.log.tpi.x.status<- ggplot(aes(x=log.tpi,y=maxn,colour=status), data=dat.ta) +
  ylab(" ")+
  xlab('Log TPI')+
  scale_color_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  #geom_jitter(width = 0.25,height = 0,alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.ta.log.tpi.x.status,show.legend=FALSE)+
  geom_line(data=predicts.ta.log.tpi.x.status,aes(y=maxn - se.fit),linetype="dashed",show.legend=FALSE)+
  geom_line(data=predicts.ta.log.tpi.x.status,aes(y=maxn + se.fit),linetype="dashed",show.legend=FALSE)+
  theme_classic()+
  Theme1#+
  #annotate("text", x = -Inf, y=Inf, label = "(b)",vjust = 1, hjust = -.1,size=5)
ggmod.ta.log.tpi.x.status

# mean.relief.by.status ----
ggmod.ta.mean.relief.x.status<- ggplot(aes(x=mean.relief,y=maxn,colour=status), data=dat.ta) +
  ylab(" ")+
  xlab('Mean relief')+
  scale_color_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  #geom_jitter(width = 0.25,height = 0,alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.ta.mean.relief.x.status,show.legend=FALSE)+
  geom_line(data=predicts.ta.mean.relief.x.status,aes(y=maxn - se.fit),linetype="dashed",show.legend=FALSE)+
  geom_line(data=predicts.ta.mean.relief.x.status,aes(y=maxn + se.fit),linetype="dashed",show.legend=FALSE)+
  theme_classic()+
  Theme1#+
#annotate("text", x = -Inf, y=Inf, label = "(b)",vjust = 1, hjust = -.1,size=5)
ggmod.ta.mean.relief.x.status

# MODEL Species richness (mean.relief+sd.relief) ----
dat.sr <- dat %>% filter(scientific=="species.richness")

gamm=gam(maxn~s(mean.relief,k=3,bs='cr')+s(sd.relief,k=3,bs='cr')+ s(site,bs="re"), family=poisson,data=dat.sr)

# predict - mean.relief ----
mod<-gamm
testdata <- expand.grid(mean.relief=seq(min(dat$mean.relief),max(dat$mean.relief),length.out = 20),
                        sd.relief=mean(mod$model$sd.relief),
                        site=(mod$model$site))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.sr.mean.relief = testdata%>%data.frame(fits)%>%
  group_by(mean.relief)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# predict - sd.relief ----
mod<-gamm
testdata <- expand.grid(sd.relief=seq(min(dat$sd.relief),max(dat$sd.relief),length.out = 20),
                        mean.relief=mean(mod$model$mean.relief),
                        site=(mod$model$site))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.sr.sd.relief = testdata%>%data.frame(fits)%>%
  group_by(sd.relief)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# PLOTS for Species richness ----
# mean relief ----
ggmod.sr.mean.relief<- ggplot() +
  ylab("Species richness")+
  xlab("Mean relief")+
  #scale_color_manual(labels = c("Fished", "SZ"),values=c("red", "black"))+
  geom_point(data=dat.sr,aes(x=mean.relief,y=maxn),  alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.sr.mean.relief,aes(x=mean.relief,y=maxn),alpha=0.5)+
  geom_line(data=predicts.sr.mean.relief,aes(x=mean.relief,y=maxn - se.fit),linetype="dashed",alpha=0.5)+
  geom_line(data=predicts.sr.mean.relief,aes(x=mean.relief,y=maxn + se.fit),linetype="dashed",alpha=0.5)+
  theme_classic()+
  Theme1+
  #annotate("text", x = -Inf, y=Inf, label = "(d)",vjust = 1, hjust = -.1,size=5)+
  annotate("text", x = -Inf, y=Inf, label = "Species richness",vjust = 1, hjust = -.1,size=5,fontface="italic")#+
  #geom_blank(data=dat.bms,aes(x=lobster,y=response*1.05))#to nudge data off annotations
ggmod.sr.mean.relief

# sd.relief ----
ggmod.sr.sd.relief<- ggplot() +
  ylab("Species richness")+
  xlab("SD relief")+
  #scale_color_manual(labels = c("Fished", "SZ"),values=c("red", "black"))+
  geom_point(data=dat.sr,aes(x=sd.relief,y=maxn),  alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.sr.sd.relief,aes(x=sd.relief,y=maxn),alpha=0.5)+
  geom_line(data=predicts.sr.sd.relief,aes(x=sd.relief,y=maxn - se.fit),linetype="dashed",alpha=0.5)+
  geom_line(data=predicts.sr.sd.relief,aes(x=sd.relief,y=maxn + se.fit),linetype="dashed",alpha=0.5)+
  theme_classic()+
  Theme1+
  #annotate("text", x = -Inf, y=Inf, label = "(d)",vjust = 1, hjust = -.1,size=5)+
  annotate("text", x = -Inf, y=Inf, label = " ",vjust = 1, hjust = -.1,size=5,fontface="italic")#+
#geom_blank(data=dat.bms,aes(x=lobster,y=response*1.05))#to nudge data off annotations
ggmod.sr.sd.relief

# MODEL Pink snapper (broad.macroalgae.by.status+log.slope+status) ----
dat.ps <- dat %>% filter(scientific=="Sparidae Chrysophrys auratus")

gamm=gam(maxn~s(broad.macroalgae,k=3,bs='cr', by=status)+s(log.slope,k=3,bs='cr')+ s(site,bs="re") + status, family=poisson,data=dat.ps)

# predict - status ----
mod<-gamm
testdata <- expand.grid(broad.macroalgae=mean(mod$model$broad.macroalgae),
                        log.slope=mean(mod$model$log.slope),
                        site=(mod$model$site),
                        status = c("Fished","No-take"))%>%
  distinct()%>%
  glimpse()


fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.ps.status = testdata%>%data.frame(fits)%>%
  group_by(status)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# predict - broad.macroalgae.by.status ----
mod<-gamm
testdata <- expand.grid(broad.macroalgae=seq(min(dat$broad.macroalgae),max(dat$broad.macroalgae),length.out = 20),
                        log.slope=mean(mod$model$log.slope),
                        site=(mod$model$site),
                        status = c("Fished","No-take"))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.ps.broad.macroalgae.x.status = testdata%>%data.frame(fits)%>%
  group_by(broad.macroalgae,status)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# predict - log.slope ----
mod<-gamm
testdata <- expand.grid(log.slope=seq(min(dat$log.slope),max(dat$log.slope),length.out = 20),
                        broad.macroalgae=mean(mod$model$broad.macroalgae),
                        site=(mod$model$site),
                        status = c("Fished","No-take"))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.ps.log.slope = testdata%>%data.frame(fits)%>%
  group_by(log.slope)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# PLOTS for Pink snapper ----
# status ----
ggmod.ps.status<- ggplot(aes(x=status,y=maxn,fill=status,colour=status), data=predicts.ps.status) +
  ylab(" ")+
  xlab('Status')+
  scale_fill_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  scale_colour_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  scale_x_discrete(limits = rev(levels(predicts.ps.status$status)))+
  geom_bar(stat = "identity")+
  geom_errorbar(aes(ymin = maxn-se.fit,ymax = maxn+se.fit),width = 0.5) +
  theme_classic()+
  Theme1+
  annotate("text", x = -Inf, y=Inf, label = "C. auratus",vjust = 1, hjust = -.1,size=5,fontface="italic")
ggmod.ps.status

# broad.macroalgae.by.status ----
ggmod.ps.broad.macroalgae.x.status<- ggplot(aes(x=broad.macroalgae,y=maxn,colour=status), data=dat.ps) +
  ylab(" ")+
  xlab('% Macroalgae')+
  scale_color_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  geom_jitter(width = 0.25,height = 0,alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.ps.broad.macroalgae.x.status,show.legend=FALSE)+
  geom_line(data=predicts.ps.broad.macroalgae.x.status,aes(y=maxn - se.fit),linetype="dashed",show.legend=FALSE)+
  geom_line(data=predicts.ps.broad.macroalgae.x.status,aes(y=maxn + se.fit),linetype="dashed",show.legend=FALSE)+
  theme_classic()+
  Theme1#+
#annotate("text", x = -Inf, y=Inf, label = "(b)",vjust = 1, hjust = -.1,size=5)
ggmod.ps.broad.macroalgae.x.status

# log.slope ----
ggmod.ps.log.slope<- ggplot() +
  ylab("")+
  xlab("Log Slope")+
  #scale_color_manual(labels = c("Fished", "SZ"),values=c("red", "black"))+
  geom_point(data=dat.ps,aes(x=log.slope,y=maxn),  alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.ps.log.slope,aes(x=log.slope,y=maxn),alpha=0.5)+
  geom_line(data=predicts.ps.log.slope,aes(x=log.slope,y=maxn - se.fit),linetype="dashed",alpha=0.5)+
  geom_line(data=predicts.ps.log.slope,aes(x=log.slope,y=maxn + se.fit),linetype="dashed",alpha=0.5)+
  theme_classic()+
  Theme1#+
  #annotate("text", x = -Inf, y=Inf, label = "(d)",vjust = 1, hjust = -.1,size=5)+
  #annotate("text", x = -Inf, y=Inf, label = "Species richness",vjust = 1, hjust = -.1,size=5,fontface="italic")#+
#geom_blank(data=dat.bms,aes(x=lobster,y=response*1.05))#to nudge data off annotations
ggmod.ps.log.slope





# MODEL Western King Wrasse (depth.by.status+sd.relief.by.status+status) ----
dat.wkw <- dat %>% filter(scientific=="Labridae Coris auricularis")

gamm=gam(maxn~s(depth,k=3,bs='cr', by=status)+s(sd.relief,k=3,bs='cr', by=status)+ s(site,bs="re") + status, family=poisson,data=dat.wkw)

# predict - status ----
mod<-gamm
testdata <- expand.grid(sd.relief=mean(mod$model$sd.relief),
                        depth=mean(mod$model$depth),
                        site=(mod$model$site),
                        status = c("Fished","No-take"))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.wkw.status = testdata%>%data.frame(fits)%>%
  group_by(status)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# predict - depth.by.status ----
mod<-gamm
testdata <- expand.grid(depth=seq(min(dat$depth),max(dat$depth),length.out = 20),
                        sd.relief=mean(mod$model$sd.relief),
                        site=(mod$model$site),
                        status = c("Fished","No-take"))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.wkw.depth.by.status = testdata%>%data.frame(fits)%>%
  group_by(depth,status)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# predict - sd.relief.by.status ----
mod<-gamm
testdata <- expand.grid(sd.relief=seq(min(dat$sd.relief),max(dat$sd.relief),length.out = 20),
                        depth=mean(mod$model$depth),
                        site=(mod$model$site),
                        status = c("Fished","No-take"))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.wkw.sd.relief.x.status = testdata%>%data.frame(fits)%>%
  group_by(sd.relief,status)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# PLOTs for Western King Wrasse ----
# status ----
ggmod.wkw.status<- ggplot(aes(x=status,y=maxn,fill=status,colour=status), data=predicts.wkw.status) +
  ylab(" ")+
  xlab('Status')+
  scale_fill_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  scale_colour_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  scale_x_discrete(limits = rev(levels(predicts.wkw.status$status)))+
  geom_bar(stat = "identity")+
  geom_errorbar(aes(ymin = maxn-se.fit,ymax = maxn+se.fit),width = 0.5) +
  theme_classic()+
  Theme1+
  annotate("text", x = -Inf, y=Inf, label = "C. auricularis",vjust = 1, hjust = -.1,size=5,fontface="italic")
ggmod.wkw.status

# depth.by.status ----
ggmod.wkw.depth.x.status<- ggplot(aes(x=depth,y=maxn,colour=status), data=dat.wkw) +
  ylab(" ")+
  xlab('Depth')+
  scale_color_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  geom_jitter(width = 0.25,height = 0,alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.wkw.depth.by.status,show.legend=FALSE)+
  geom_line(data=predicts.wkw.depth.by.status,aes(y=maxn - se.fit),linetype="dashed",show.legend=FALSE)+
  geom_line(data=predicts.wkw.depth.by.status,aes(y=maxn + se.fit),linetype="dashed",show.legend=FALSE)+
  theme_classic()+
  Theme1#+
#annotate("text", x = -Inf, y=Inf, label = "(b)",vjust = 1, hjust = -.1,size=5)
ggmod.wkw.depth.x.status

# sd.relief.by.status ----
ggmod.wkw.sd.relief.x.status<- ggplot(aes(x=sd.relief,y=maxn,colour=status), data=dat.wkw) +
  ylab(" ")+
  xlab('SD relief')+
  scale_color_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  geom_jitter(width = 0.25,height = 0,alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.wkw.sd.relief.x.status,show.legend=FALSE)+
  geom_line(data=predicts.wkw.sd.relief.x.status,aes(y=maxn - se.fit),linetype="dashed",show.legend=FALSE)+
  geom_line(data=predicts.wkw.sd.relief.x.status,aes(y=maxn + se.fit),linetype="dashed",show.legend=FALSE)+
  theme_classic()+
  Theme1
ggmod.wkw.sd.relief.x.status

# MODEL Port Jackson Shark (distance.to.ramp+log.slope+mean.relief) ----
dat.pj <- dat %>% filter(scientific=="Heterodontidae Heterodontus portusjacksoni")

gamm=gam(maxn~s(distance.to.ramp,k=3,bs='cr')+s(log.slope,k=3,bs='cr')+s(mean.relief,k=3,bs='cr')+ s(site,bs="re"), family=poisson,data=dat.pj)

# predict - mean.relief ----
mod<-gamm
testdata <- expand.grid(mean.relief=seq(min(dat$mean.relief),max(dat$mean.relief),length.out = 20),
                        log.slope=mean(mod$model$log.slope),
                        distance.to.ramp=mean(mod$model$distance.to.ramp),
                        site=(mod$model$site))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.pj.mean.relief = testdata%>%data.frame(fits)%>%
  group_by(mean.relief)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# predict - distance.to.ramp ----
mod<-gamm
testdata <- expand.grid(distance.to.ramp=seq(min(dat$distance.to.ramp),max(dat$distance.to.ramp),length.out = 20),
                        mean.relief=mean(mod$model$mean.relief),
                        log.slope=mean(mod$model$log.slope),
                        site=(mod$model$site))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.pj.distance.to.ramp = testdata%>%data.frame(fits)%>%
  group_by(distance.to.ramp)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# predict - log.slope ----
mod<-gamm
testdata <- expand.grid(log.slope=seq(min(dat$log.slope),max(dat$log.slope),length.out = 20),
                        mean.relief=mean(mod$model$mean.relief),
                        distance.to.ramp=mean(mod$model$distance.to.ramp),
                        site=(mod$model$site))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.pj.log.slope = testdata%>%data.frame(fits)%>%
  group_by(log.slope)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# Plots for Port Jackson Shark ----
# log.slope ----
ggmod.pj.log.slope<- ggplot() +
  ylab("")+
  xlab("Log Slope")+
  #scale_color_manual(labels = c("Fished", "SZ"),values=c("red", "black"))+
  geom_point(data=dat.pj,aes(x=log.slope,y=maxn),  alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.pj.log.slope,aes(x=log.slope,y=maxn),alpha=0.5)+
  geom_line(data=predicts.pj.log.slope,aes(x=log.slope,y=maxn - se.fit),linetype="dashed",alpha=0.5)+
  geom_line(data=predicts.pj.log.slope,aes(x=log.slope,y=maxn + se.fit),linetype="dashed",alpha=0.5)+
  theme_classic()+
  Theme1#+
#annotate("text", x = -Inf, y=Inf, label = "(d)",vjust = 1, hjust = -.1,size=5)+
#annotate("text", x = -Inf, y=Inf, label = "Species richness",vjust = 1, hjust = -.1,size=5,fontface="italic")#+
#geom_blank(data=dat.bms,aes(x=lobster,y=response*1.05))#to nudge data off annotations
ggmod.pj.log.slope

# distance.to.ramp ---- 
ggmod.pj.distance.to.ramp <- ggplot() +
  ylab("")+
  xlab("Distance to ramp")+
  #scale_color_manual(labels = c("Fished", "SZ"),values=c("red", "black"))+
  geom_point(data=dat.pj,aes(x=distance.to.ramp,y=maxn),  alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.pj.distance.to.ramp,aes(x=distance.to.ramp,y=maxn),alpha=0.5)+
  geom_line(data=predicts.pj.distance.to.ramp,aes(x=distance.to.ramp,y=maxn - se.fit),linetype="dashed",alpha=0.5)+
  geom_line(data=predicts.pj.distance.to.ramp,aes(x=distance.to.ramp,y=maxn + se.fit),linetype="dashed",alpha=0.5)+
  theme_classic()+
  Theme1#+
#annotate("text", x = -Inf, y=Inf, label = "(d)",vjust = 1, hjust = -.1,size=5)+
#annotate("text", x = -Inf, y=Inf, label = "Species richness",vjust = 1, hjust = -.1,size=5,fontface="italic")#+
#geom_blank(data=dat.bms,aes(x=lobster,y=response*1.05))#to nudge data off annotations
ggmod.pj.distance.to.ramp

# mean.relief ---- 
ggmod.pj.mean.relief <- ggplot() +
  ylab("")+
  xlab("Mean relief")+
  #scale_color_manual(labels = c("Fished", "SZ"),values=c("red", "black"))+
  geom_point(data=dat.pj,aes(x=mean.relief,y=maxn),  alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.pj.mean.relief,aes(x=mean.relief,y=maxn),alpha=0.5)+
  geom_line(data=predicts.pj.mean.relief,aes(x=mean.relief,y=maxn - se.fit),linetype="dashed",alpha=0.5)+
  geom_line(data=predicts.pj.mean.relief,aes(x=mean.relief,y=maxn + se.fit),linetype="dashed",alpha=0.5)+
  theme_classic()+
  Theme1#+
#annotate("text", x = -Inf, y=Inf, label = "(d)",vjust = 1, hjust = -.1,size=5)+
#annotate("text", x = -Inf, y=Inf, label = "Species richness",vjust = 1, hjust = -.1,size=5,fontface="italic")#+
#geom_blank(data=dat.bms,aes(x=lobster,y=response*1.05))#to nudge data off annotations
ggmod.pj.mean.relief


# INSIDE/OUTSIDE -----
dat <- maxn.io

unique(dat$scientific)

# MODEL Total abundance (log.tpi.by.status+mean.relief.by.status+status) ----
dat.ta <- dat %>% filter(scientific=="total.abundance")

gamm=gam(maxn~s(log.tpi,k=3,bs='cr', by=status)+s(mean.relief,k=3,bs='cr', by=status)+ s(site,bs="re") + status, family=poisson,data=dat.ta)

# predict - status ----
mod<-gamm
testdata <- expand.grid(mean.relief=mean(mod$model$mean.relief),
                        log.tpi=mean(mod$model$log.tpi),
                        site=(mod$model$site),
                        status = c("Fished","No-take"))%>%
  distinct()%>%
  glimpse()


fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.ta.status = testdata%>%data.frame(fits)%>%
  group_by(status)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# predict - log.tpi.by.status ----
mod<-gamm
testdata <- expand.grid(log.tpi=seq(min(dat$log.tpi),max(dat$log.tpi),length.out = 20),
                        mean.relief=mean(mod$model$mean.relief),
                        site=(mod$model$site),
                        status = c("Fished","No-take"))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.ta.log.tpi.x.status = testdata%>%data.frame(fits)%>%
  group_by(log.tpi,status)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# predict - mean.relief.by.status ----
mod<-gamm
testdata <- expand.grid(mean.relief=seq(min(dat$mean.relief),max(dat$mean.relief),length.out = 20),
                        log.tpi=mean(mod$model$log.tpi),
                        site=(mod$model$site),
                        status = c("Fished","No-take"))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.ta.mean.relief.x.status = testdata%>%data.frame(fits)%>%
  group_by(mean.relief,status)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()


# PLOTS for Total abundance ----
# status ----
ggmod.ta.status.io <- ggplot(aes(x=status,y=maxn,fill=status,colour=status), data=predicts.ta.status) +
  ylab(" ")+
  xlab('Status')+
  scale_fill_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  scale_colour_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  scale_x_discrete(limits = rev(levels(predicts.ta.status$status)))+
  geom_bar(stat = "identity")+
  geom_errorbar(aes(ymin = maxn-se.fit,ymax = maxn+se.fit),width = 0.5) +
  theme_classic()+
  Theme1+
  annotate("text", x = -Inf, y=Inf, label = "Total abundance",vjust = 1, hjust = -.1,size=5,fontface="italic")
ggmod.ta.status.io

# log.tpi.by.status ----
ggmod.ta.log.tpi.x.status.io <- ggplot(aes(x=log.tpi,y=maxn,colour=status), data=dat.ta) +
  ylab(" ")+
  xlab('Log TPI')+
  scale_color_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  geom_jitter(width = 0.25,height = 0,alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.ta.log.tpi.x.status,show.legend=FALSE)+
  geom_line(data=predicts.ta.log.tpi.x.status,aes(y=maxn - se.fit),linetype="dashed",show.legend=FALSE)+
  geom_line(data=predicts.ta.log.tpi.x.status,aes(y=maxn + se.fit),linetype="dashed",show.legend=FALSE)+
  theme_classic()+
  Theme1#+
#annotate("text", x = -Inf, y=Inf, label = "(b)",vjust = 1, hjust = -.1,size=5)
ggmod.ta.log.tpi.x.status.io

# mean.relief.by.status ----
ggmod.ta.mean.relief.x.status.io <- ggplot(aes(x=mean.relief,y=maxn,colour=status), data=dat.ta) +
  ylab(" ")+
  xlab('Mean relief')+
  scale_color_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  geom_jitter(width = 0.25,height = 0,alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.ta.mean.relief.x.status,show.legend=FALSE)+
  geom_line(data=predicts.ta.mean.relief.x.status,aes(y=maxn - se.fit),linetype="dashed",show.legend=FALSE)+
  geom_line(data=predicts.ta.mean.relief.x.status,aes(y=maxn + se.fit),linetype="dashed",show.legend=FALSE)+
  theme_classic()+
  Theme1#+
#annotate("text", x = -Inf, y=Inf, label = "(b)",vjust = 1, hjust = -.1,size=5)
ggmod.ta.mean.relief.x.status.io


# MODEL Species richness (mean.relief) ----
dat.sr <- dat %>% filter(scientific=="species.richness")

gamm=gam(maxn~s(mean.relief,k=3,bs='cr')+ s(site,bs="re"), family=poisson,data=dat.sr)

# predict - mean.relief ----
mod<-gamm
testdata <- expand.grid(mean.relief=seq(min(dat$mean.relief),max(dat$mean.relief),length.out = 20),
                        site=(mod$model$site))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.sr.mean.relief = testdata%>%data.frame(fits)%>%
  group_by(mean.relief)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()


# PLOTS for Species richness ----
# mean relief ----
ggmod.sr.mean.relief.io <- ggplot() +
  ylab("Species richness")+
  xlab("Mean relief")+
  #scale_color_manual(labels = c("Fished", "SZ"),values=c("red", "black"))+
  geom_point(data=dat.sr,aes(x=mean.relief,y=maxn),  alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.sr.mean.relief,aes(x=mean.relief,y=maxn),alpha=0.5)+
  geom_line(data=predicts.sr.mean.relief,aes(x=mean.relief,y=maxn - se.fit),linetype="dashed",alpha=0.5)+
  geom_line(data=predicts.sr.mean.relief,aes(x=mean.relief,y=maxn + se.fit),linetype="dashed",alpha=0.5)+
  theme_classic()+
  Theme1+
  #annotate("text", x = -Inf, y=Inf, label = "(d)",vjust = 1, hjust = -.1,size=5)+
  annotate("text", x = -Inf, y=Inf, label = "Species richness",vjust = 1, hjust = -.1,size=5,fontface="italic")#+
#geom_blank(data=dat.bms,aes(x=lobster,y=response*1.05))#to nudge data off annotations
ggmod.sr.mean.relief.io

# MODEL Pink snapper (log.slope.by.status+mean.relief+status) ----
dat.ps <- dat %>% filter(scientific=="Sparidae Chrysophrys auratus")

gamm=gam(maxn~s(log.slope,k=3,bs='cr', by=status)+s(mean.relief,k=3,bs='cr')+ s(site,bs="re") + status, family=poisson,data=dat.ps)

# predict - status ----
mod<-gamm
testdata <- expand.grid(mean.relief=mean(mod$model$mean.relief),
                        log.slope=mean(mod$model$log.slope),
                        site=(mod$model$site),
                        status = c("Fished","No-take"))%>%
  distinct()%>%
  glimpse()


fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.ps.status = testdata%>%data.frame(fits)%>%
  group_by(status)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# predict - log.slope ----
mod<-gamm
testdata <- expand.grid(log.slope=seq(min(dat$log.slope),max(dat$log.slope),length.out = 20),
                        mean.relief=mean(mod$model$mean.relief),
                        site=(mod$model$site),
                        status = c("Fished","No-take"))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.ps.log.slope.x.status = testdata%>%data.frame(fits)%>%
  group_by(log.slope,status)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# predict - mean.relief ----
mod<-gamm
testdata <- expand.grid(mean.relief=seq(min(dat$mean.relief),max(dat$mean.relief),length.out = 20),
                        log.slope=mean(mod$model$log.slope),
                        site=(mod$model$site),
                        status = c("Fished","No-take"))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.ps.mean.relief = testdata%>%data.frame(fits)%>%
  group_by(mean.relief)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# PLOTS for Pink snapper ----
# status ----
ggmod.ps.status.io <- ggplot(aes(x=status,y=maxn,fill=status,colour=status), data=predicts.ps.status) +
  ylab(" ")+
  xlab('Status')+
  scale_fill_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  scale_colour_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  scale_x_discrete(limits = rev(levels(predicts.ps.status$status)))+
  geom_bar(stat = "identity")+
  geom_errorbar(aes(ymin = maxn-se.fit,ymax = maxn+se.fit),width = 0.5) +
  theme_classic()+
  Theme1+
  annotate("text", x = -Inf, y=Inf, label = "C. auratus",vjust = 1, hjust = -.1,size=5,fontface="italic")
ggmod.ps.status.io

# log.slope.by.status ----
ggmod.ps.log.slope.x.status.io <- ggplot(aes(x=log.slope,y=maxn,colour=status), data=dat.ps) +
  ylab(" ")+
  xlab('log.slope')+
  scale_color_manual(labels = c("Fished", "No-take"),values=c("red", "black"))+
  geom_jitter(width = 0.25,height = 0,alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.ps.log.slope.x.status,show.legend=FALSE)+
  geom_line(data=predicts.ps.log.slope.x.status,aes(y=maxn - se.fit),linetype="dashed",show.legend=FALSE)+
  geom_line(data=predicts.ps.log.slope.x.status,aes(y=maxn + se.fit),linetype="dashed",show.legend=FALSE)+
  theme_classic()+
  Theme1#+
#annotate("text", x = -Inf, y=Inf, label = "(b)",vjust = 1, hjust = -.1,size=5)
ggmod.ps.log.slope.x.status.io

# mean.relief ----
ggmod.ps.mean.relief.io <- ggplot() +
  ylab("")+
  xlab("mean.relief")+
  #scale_color_manual(labels = c("Fished", "SZ"),values=c("red", "black"))+
  geom_point(data=dat.ps,aes(x=mean.relief,y=maxn),  alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.ps.mean.relief,aes(x=mean.relief,y=maxn),alpha=0.5)+
  geom_line(data=predicts.ps.mean.relief,aes(x=mean.relief,y=maxn - se.fit),linetype="dashed",alpha=0.5)+
  geom_line(data=predicts.ps.mean.relief,aes(x=mean.relief,y=maxn + se.fit),linetype="dashed",alpha=0.5)+
  theme_classic()+
  Theme1#+
#annotate("text", x = -Inf, y=Inf, label = "(d)",vjust = 1, hjust = -.1,size=5)+
#annotate("text", x = -Inf, y=Inf, label = "Species richness",vjust = 1, hjust = -.1,size=5,fontface="italic")#+
#geom_blank(data=dat.bms,aes(x=lobster,y=response*1.05))#to nudge data off annotations
ggmod.ps.mean.relief.io











# MODEL Port Jackson Shark (depth+log.roughness) ----
dat.pj <- dat %>% filter(scientific=="Heterodontidae Heterodontus portusjacksoni")

gamm=gam(maxn~s(depth,k=3,bs='cr')+s(log.roughness,k=3,bs='cr')+s(site,bs="re"), family=poisson,data=dat.pj)

# predict - depth ----
mod<-gamm
testdata <- expand.grid(depth=seq(min(dat$depth),max(dat$depth),length.out = 20),
                        log.roughness=mean(mod$model$log.roughness),
                        site=(mod$model$site))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.pj.depth = testdata%>%data.frame(fits)%>%
  group_by(depth)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()

# predict - log.roughness ----
mod<-gamm
testdata <- expand.grid(log.roughness=seq(min(dat$log.roughness),max(dat$log.roughness),length.out = 20),
                        depth=mean(mod$model$depth),
                        site=(mod$model$site))%>%
  distinct()%>%
  glimpse()

fits <- predict.gam(mod, newdata=testdata, type='response', se.fit=T)

predicts.pj.log.roughness = testdata%>%data.frame(fits)%>%
  group_by(log.roughness)%>% #only change here
  summarise(maxn=mean(fit),se.fit=mean(se.fit))%>%
  ungroup()


# Plots for Port Jackson Shark ----
# depth ----
ggmod.pj.depth.io<- ggplot() +
  ylab("")+
  xlab("Depth")+
  #scale_color_manual(labels = c("Fished", "SZ"),values=c("red", "black"))+
  geom_point(data=dat.pj,aes(x=depth,y=maxn),  alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.pj.depth,aes(x=depth,y=maxn),alpha=0.5)+
  geom_line(data=predicts.pj.depth,aes(x=depth,y=maxn - se.fit),linetype="dashed",alpha=0.5)+
  geom_line(data=predicts.pj.depth,aes(x=depth,y=maxn + se.fit),linetype="dashed",alpha=0.5)+
  theme_classic()+
  Theme1#+
#annotate("text", x = -Inf, y=Inf, label = "(d)",vjust = 1, hjust = -.1,size=5)+
#annotate("text", x = -Inf, y=Inf, label = "Species richness",vjust = 1, hjust = -.1,size=5,fontface="italic")#+
#geom_blank(data=dat.bms,aes(x=lobster,y=response*1.05))#to nudge data off annotations
ggmod.pj.depth.io

# distance.to.ramp ---- 
ggmod.pj.distance.to.ramp <- ggplot() +
  ylab("")+
  xlab("Distance to ramp")+
  #scale_color_manual(labels = c("Fished", "SZ"),values=c("red", "black"))+
  geom_point(data=dat.pj,aes(x=distance.to.ramp,y=maxn),  alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.pj.distance.to.ramp,aes(x=distance.to.ramp,y=maxn),alpha=0.5)+
  geom_line(data=predicts.pj.distance.to.ramp,aes(x=distance.to.ramp,y=maxn - se.fit),linetype="dashed",alpha=0.5)+
  geom_line(data=predicts.pj.distance.to.ramp,aes(x=distance.to.ramp,y=maxn + se.fit),linetype="dashed",alpha=0.5)+
  theme_classic()+
  Theme1#+
#annotate("text", x = -Inf, y=Inf, label = "(d)",vjust = 1, hjust = -.1,size=5)+
#annotate("text", x = -Inf, y=Inf, label = "Species richness",vjust = 1, hjust = -.1,size=5,fontface="italic")#+
#geom_blank(data=dat.bms,aes(x=lobster,y=response*1.05))#to nudge data off annotations
ggmod.pj.distance.to.ramp

# mean.relief ---- 
ggmod.pj.mean.relief <- ggplot() +
  ylab("")+
  xlab("Mean relief")+
  #scale_color_manual(labels = c("Fished", "SZ"),values=c("red", "black"))+
  geom_point(data=dat.pj,aes(x=mean.relief,y=maxn),  alpha=0.75, size=2,show.legend=FALSE)+
  geom_line(data=predicts.pj.mean.relief,aes(x=mean.relief,y=maxn),alpha=0.5)+
  geom_line(data=predicts.pj.mean.relief,aes(x=mean.relief,y=maxn - se.fit),linetype="dashed",alpha=0.5)+
  geom_line(data=predicts.pj.mean.relief,aes(x=mean.relief,y=maxn + se.fit),linetype="dashed",alpha=0.5)+
  theme_classic()+
  Theme1#+
#annotate("text", x = -Inf, y=Inf, label = "(d)",vjust = 1, hjust = -.1,size=5)+
#annotate("text", x = -Inf, y=Inf, label = "Species richness",vjust = 1, hjust = -.1,size=5,fontface="italic")#+
#geom_blank(data=dat.bms,aes(x=lobster,y=response*1.05))#to nudge data off annotations
ggmod.pj.mean.relief



# combined.plot using grid() and gridExtra()------
blank <- grid.rect(gp=gpar(col="white"))

# To see what they will look like use grid.arrange() - make sure Plot window is large enough! - or will error!
grid.arrange(ggmod.bds.status,ggmod.bds.distance.x.status,ggmod.bds.500um,
             ggmod.bms.lobster,blank,blank,
             ggmod.cpn.lobster,ggmod.cpn.4mm,blank,nrow=3,ncol=3)

# Use arrangeGrob ONLY - as we can pass this to ggsave! Note use of raw ggplot's
combine.plot<-arrangeGrob(ggmod.bds.status,ggmod.bds.distance.x.status,ggmod.bds.500um,
                          ggmod.bms.lobster,blank,blank,
                          ggmod.cpn.lobster,ggmod.cpn.4mm,blank,nrow=3,ncol=3)

ggsave(combine.plot,file="Langlois_gamm.plot.png", width = 30, height = 30,units = "cm")
